golang/aoc2023/day05/second_test.go
package day05

import (
	"strings"
	"syodage/aoc2023/utils"
	"testing"
)

func TestSecondPart(t *testing.T) {

	secondInput, err := utils.ReadLines("./second_input.txt")
	if err != nil {
		t.Errorf("Error reading input file")
	}

	tests := []struct {
		name  string
		input []string
		want  int
	}{
		{
			name:  "SampleInput",
			input: strings.Split(firstSampleTest, "\n"),
			want:  46,
		},
		{
			name:  "PuzzleInput",
			input: secondInput,
			want:  20191102,
		},
	}

	for _, test := range tests {
		t.Run(test.name, func(t *testing.T) {
			got, err := SecondPart(test.input)
			if err != nil {
				t.Errorf("Returned error: %v", err)
			}

			if got != test.want {
				t.Errorf("Want: %d, got: %d", test.want, got)
			}
		})
	}
}
